This is the deflation sequences we obtained.
This is the deflation sequences we obtained for B(3,3,3|23). The Defsequence is a table of form: 17376×4. 
Each row is the first 4 numbers (n0,n1,n2,n3) of the deflation sequence of a solution of the Brent equation of type [3,3,3|23]. All the solutions are obtained from the database: http://www.algebra.uni-linz.ac.at/research/matrix-multiplication/

We order them in the form of Equation (5.3) of our arxiv paper, which can be seen in the data: Solution333  (a table of format: 621×17376).

The name of each solution is the data: Name.