Deflation sequence for B(4,4,4|49)
This is the deflation sequences we obtained for B(4,4,4|49).
The Defsequence is a table of form: 14236×2. Each row is the first 2 numbers (n0,n1) of the deflation sequence of a solution of the Brent equation of type [4,4,4|49]. All the solutions are obtained from the database: https://github.com/deepmind/alphatensor/
We order them in the form of Equation (5.3) of our arxiv paper, which can be seen in the data: Solution444  (a table of format: 2352×14236).
The name of each solution is the data: Name.